// Web/RMA-Web-Shell/projects/medicare/src/app/medical-invoice-manager/models/medical-switch-batch-invoice.ts
import { MedicalSwitchBatchInvoiceLine } from 'projects/medicare/src/app/medical-invoice-manager/models/medical-switch-batch-invoice-line';
import { BaseClass } from 'projects/shared-models-lib/src/lib/common/base-class';
import { SwitchInvoiceStatusEnum } from '../enums/switch-invoice-status-enum';
import { InvoiceUnderAssessReason } from './invoice-under-assess-reason';
import { SwitchBatchInvoiceUnderAssessReason } from './switch-batch-invoice-under-assess-reason';
import { SwitchBatchType } from '../../shared/enums/switch-batch-type';

export class MedicalSwitchBatchInvoice extends BaseClass  {
    switchBatchInvoiceId: number;
    switchBatchId: number;
    invoiceId: number | null;
    batchSequenceNumber: number | null;
    switchBatchNumber: string;
    switchTransactionNumber: string;
    practiceNumber: string;
    healthCareProviderName: string;
    invoiceDate: string | null;
    dateSubmitted: string | null;
    dateReceived: string | null;
    spInvoiceNumber: string;
    spAccountNumber: string;
    totalInvoiceAmount: number | null;
    totalInvoiceVat: number | null;
    totalInvoiceAmountInclusive: number | null;
    dateAdmitted: string | null;
    dateDischarged: string | null;
    claimReferenceNumber: string;
    eventDate: string | null;
    surname: string;
    firstName: string;
    initials: string;
    dateOfBirth: string | null;
    idNumber: string;
    companyNumber: string;
    preAuthNumber: string;
    possiblePersonEventId: number | null;
    possibleEventId: number | null;
    healthCareProviderId: number | null;
    preAuthId: number | null;
    isProcessed: boolean | null;
    treatingDocBhf: string;
    employerName: string;
    referringDocBhf: string;
    referredTo: string;
    hospitalIndicator: string;
    surgeonBhfNumber: string;
    anaesthetistBhfNumber: string;
    assistantBhfNumber: string;
    lengthOfStay: string;
    freeTextDiagnosis: string;
    iodReference: string;
    patientGender: string;
    diagnosticCodeType: string;
    disciplineCode: string;
    claimId : number | null;
    claimReferenceNumberMatch : string;
    switchBatchInvoiceLines: MedicalSwitchBatchInvoiceLine[];
    activeUnderAssessReason: string;
    switchInvoiceStatus: SwitchInvoiceStatusEnum;
    status: string;
    employeeNumber: string;
    reinstateReason: string;
    medicalInvoiceUnderAssessReasons: InvoiceUnderAssessReason[];
    switchBatchInvoiceUnderAssessReasons: SwitchBatchInvoiceUnderAssessReason[];
    switchBatchType: SwitchBatchType | null;
}